requirements.txt
Flask==3.0.0
Flask-Mail==0.10.0
mysql-connector-python==8.2.0
Pillow>=10.2.0
qrcode[pil]>=7.4.2
python-dotenv>=1.0.0
opencv-python>=4.8.0
numpy>=1.26.0
PyPDF2>=3.0.1
Werkzeug>=3.0.0
gunicorn>=21.2.0
